refactor: extract ODataBatchHandler partials (#1144)
Split the 2,312-LOC ODataBatchHandler.cs into focused partials so each
file holds a single responsibility from the batch pipeline. No behavior
change; the public surface (constructor and ProcessBatchAsync) stays on
the main type.

- ODataBatchHandler.cs (232) — public surface, ProcessBatchAsync, fields,
  constants, LoggerMessage Log nested class.
- ODataBatchHandler.Dependencies.cs (418) — dependency ordering
  (HasDependencies, OrderRequestsByDependencies, DependencyOrderResult)
  and Content-ID reference resolution (TryResolveContentIdReferences and
  friends).
- ODataBatchHandler.ChangeSet.cs (785) — atomicity-group / change-set
  processing (ProcessAtomicGroupAsync).
- ODataBatchHandler.Dispatch.cs (651) — per-request dispatch
  (ProcessSingleRequestAsync), URL normalization, batch HttpContext
  creation, response item assembly from IResult, path parsing,
  TryParseCollectionQueryOptions, and access validation.
- ODataBatchHandler.Edit.cs (248) — BuildValidatedEditBatchAsync,
  CreateFeatureFromBodyAsync, ResolveAxisOrderAsync, and
  ValidatePreconditionsAsync.
- ODataBatchHandler.Response.cs (80) — FeatureToBody, ComputeFeatureEtag,
  CreateSuccessResponse / CreateErrorResponse envelope helpers.
